Move Liveblog test into separate class
